requirements.txt
python-whois
requests
beautifulsoup4
langdetect
psycopg2-binary